fix(taskmanager): create export dialog, fix bg-img, fix nav, create each components file
