+++ src/app/globals.css
@@ -11,6 +11,7 @@
     background-size: cover; /* Ensures the image covers the whole screen */
     background-repeat: no-repeat; /* Prevents repeating */
     background-position: center; /* Centers the image */
+    background-attachment: fixed; 
   }
   h1 {
     @apply font-Anuphan text-[48px] leading-[48px] font-semibold;
