+++ src/app/projects/page.tsx
@@ -3,7 +3,7 @@ import { ProjectList } from '@/components/elements/projectList';
 
 export default function ProjectLists() {
   return (
-    <div className=" flex flex-wrap flex- w-full items-start gap-[16px]  ">
+    <div className="flex flex-wrap flex- w-full items-start gap-[16px]  ">
       <ProjectList />
     </div>
   );
