+++ src/components/elements/nav-bar.tsx
@@ -1,26 +1,23 @@
 'use client';
 
-import { Profile, Profile2 } from './profile';
-import { useEffect, useState } from 'react';
+import { usePathname } from 'next/navigation';
+import { Profile2 } from './profile';
 
 export default function NavBar() {
-  const [isHidden, setIsHidden] = useState('');
-
-  useEffect(() => {
-    const url = window.location.pathname;
-    if (url === '/') {
-      setIsHidden('hidden');
-    }
-  }, []);
+  const url = usePathname();
 
   return (
-    <div className={`flex flex-row min-w-full h-[84px] justify-between items-center ${isHidden}`}>
-      <a href="/projects/cm24w5yu000008tlglutu5czu">
-        <img src="/asset/logo/Logo_s2.svg" alt="CUSW" />
-      </a>
-      <div className="flex flex-row px-5">
-        <Profile2 userId={''} userName={'Bunyawat Naunnak'} />
-      </div>
-    </div>
+    <>
+      {url !== '/' && (
+        <div className="flex flex-row min-w-full h-[84px] justify-between items-center">
+          <a href="/projects/cm24w5yu000008tlglutu5czu">
+            <img src="/asset/logo/Logo_s2.svg" alt="CUSW" />
+          </a>
+          <div className="flex flex-row px-5">
+            <Profile2 userId={''} userName={'Bunyawat Naunnak'} />
+          </div>
+        </div>
+      )}
+    </>
   );
 }
